refactor: pull SamplerFactory.java into SampleStrategy.java
